+++ CHANGELOG.md
@@ -21,6 +21,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - special for image displaying theme
 - multiple selection mode in BaseSelect
 - BaseColorPicker with more color choices, transparency changes and color model changes
+- label displaying in BaseRadio 
+- support 'max' and 'min' attributes for BaseInput
+- ability to change placement in BaseRadio
 
 ### Changed